requirements.txt
aiofiles==24.1.0
aiohttp==3.12.14
alembic==1.16.4
anytree==2.13.0
Babel==2.17.0
beautifulsoup4==4.13.4
certifi==2025.7.14
chardet==5.2.0
click==8.1.8
cryptography==45.0.5
dnspython==2.7.0
fastapi==0.116.1
fastapi-csrf-protect==1.0.3
Flask==3.1.1
freezegun==1.5.3
h11==0.16.0
idna==3.10
Jinja2==3.1.6
git+https://github.com/CERT-Polska/mailgoose@v1.3.15#egg=libmailgoose&subdirectory=scan
liccheck==0.9.2
Markdown==3.8.2
MarkupSafe==3.0.2
minio==7.2.1
more-itertools==10.7.0
openapi-spec-validator==0.7.2
packaging==25.0
paramiko==3.5.1
prometheus-client==0.22.1
psutil==7.0.0
psycopg2-binary==2.9.10
publicsuffixlist==1.0.2.20250719
pydantic==2.10.6
PyMySQL==1.1.1
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
redis==6.2.0
requests==2.32.4
requests-cache==1.2.1
requests-mock==1.12.1
retry==0.9.2
SQLAlchemy==1.4.52   # downgrade to avoid greenlet issue
semver==3.0.4
setuptools==80.9.0
shodan==1.31.0
six==1.17.0
termcolor==3.1.0
timeout-decorator==0.5.0
tqdm==4.67.1
typer==0.16.0
typing-extensions==4.14.1
urllib3==1.26.18
uvicorn==0.35.0
validators==0.35.0
whoisdomain==1.20250220.2
celery==5.3.6
karton-core
